Refactor serving code
1. Fixed soptbugs error
2. Disabled parallel test, run ConfigMangerTest will cause build flaky
3. Minor code refacor
